# server/docs/swagger.yaml
basePath: /api
definitions:
  github_com_ai-comic-generator_server_internal_common.BaseResponse:
    properties:
      code:
        description: 业务状态码，0 表示成功，非 0 表示失败
        example: 0
        type: integer
      data:
        description: 响应数据，成功时返回；失败时通常为 null
      message:
        description: 响应消息，成功时为 ok，失败时为错误描述
        example: ok
        type: string
    type: object
...
  github_com_ai-comic-generator_server_internal_model.AddUserRequest:
    properties:
      points:
        description: 积分（可选，不传则使用库表默认值）
        example: 100
        type: integer
      status:
        description: 状态（可选）
        example: 1
        type: integer
      userAccount:
        description: 登录账号
        example: newuser
        type: string
      userAvatar:
        description: 头像 URL（可选）
        example: https://example.com/avatar.png
        type: string
      userName:
        description: 用户昵称（可选）
        example: 新用户
        type: string
      userProfile:
        description: 个人简介（可选）
        example: 个人简介
        type: string
      userRole:
        description: 用户角色：user / admin
        enum:
        - user
        - admin
        example: user
        type: string
    required:
    - userAccount
    type: object
...
  github_com_ai-comic-generator_server_internal_model.DeleteRequest:
    properties:
      id:
        description: 要删除的记录主键 ID（用户或文章）
        example: 1
        type: integer
    required:
    - id
    type: object
  github_com_ai-comic-generator_server_internal_model.LoginRequest:
    properties:
      userAccount:
        description: 登录账号
        example: zhangsan
        type: string
      userPassword:
        description: 登录密码
        example: "12345678"
        type: string
    type: object
...
  github_com_ai-comic-generator_server_internal_model.QueryUserRequest:
    properties:
      id:
        description: 按用户 ID 筛选（可选）
        example: 1
        type: integer
      pageNum:
        description: 页码，从 1 开始
        example: 1
        type: integer
      pageSize:
        description: 每页条数
        example: 10
        type: integer
      sortField:
        description: 排序字段（可选）
        example: createTime
        type: string
      sortOrder:
        description: 排序方向：asc / desc（可选）
        example: desc
        type: string
      userAccount:
        description: 按账号模糊筛选（可选）
        example: zhangsan
        type: string
      userName:
        description: 按昵称模糊筛选（可选）
        example: 张三
        type: string
      userProfile:
        description: 按简介模糊筛选（可选）
        type: string
      userRole:
        description: 按角色筛选（可选）
        example: user
        type: string
    type: object
...
  github_com_ai-comic-generator_server_internal_model.RegisterRequest:
    properties:
      checkPassword:
        description: 确认密码，需与 userPassword 一致
        example: "12345678"
        minLength: 8
        type: string
      userAccount:
        description: 登录账号，至少 4 位
        example: zhangsan
        minLength: 4
        type: string
      userPassword:
        description: 登录密码，至少 8 位
        example: "12345678"
        minLength: 8
        type: string
    required:
    - checkPassword
    - userAccount
    - userPassword
    type: object
...
  github_com_ai-comic-generator_server_internal_model.UpdatePasswordRequest:
    properties:
      checkPassword:
        description: 确认新密码
        example: "87654321"
        minLength: 8
        type: string
      newPassword:
        description: 新密码
        example: "87654321"
        minLength: 8
        type: string
      oldPassword:
        description: 原密码
        example: "12345678"
        minLength: 8
        type: string
    required:
    - checkPassword
    - newPassword
    - oldPassword
    type: object
...
  github_com_ai-comic-generator_server_internal_model.UpdateProfileRequest:
    properties:
      userAvatar:
        description: 头像 URL（可选）
        example: https://example.com/avatar.png
        type: string
      userName:
        description: 用户昵称（可选）
        example: 新昵称
        type: string
      userProfile:
        description: 个人简介（可选）
        example: 个人简介
        type: string
    type: object
  github_com_ai-comic-generator_server_internal_model.UpdateUserRequest:
    properties:
      id:
        description: 用户主键 ID
        example: 1
        type: integer
      points:
        description: 积分（可选）
        example: 100
        type: integer
      status:
        description: 状态（可选）
        example: 1
        type: integer
      userAvatar:
        description: 头像 URL（可选）
        example: https://example.com/avatar.png
        type: string
      userName:
        description: 用户昵称（可选）
        example: 新昵称
        type: string
      userProfile:
        description: 个人简介（可选）
        example: 更新后的简介
        type: string
      userRole:
        description: 用户角色（可选）
        enum:
        - user
        - admin
        example: user
        type: string
    required:
    - id
    type: object
info:
  contact: {}
  description: REST API。HTTP 多为 200，业务结果看 JSON 的 code（0 成功）；鉴权接口需 Cookie session。
  title: AI Comic Generator API
  version: "1.0"
paths:
...
  /user/add:
    post:
      consumes:
      - application/json
      description: 管理员创建用户，需 admin 角色
      parameters:
      - description: 创建参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_ai-comic-generator_server_internal_model.AddUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 业务错误（无权限、参数错误等）
          schema:
            $ref: '#/definitions/github_com_ai-comic-generator_server_internal_common.BaseResponse'
      security:
      - SessionCookie: []
      summary: 创建用户
      tags:
      - 用户管理
  /user/delete:
    post:
      consumes:
      - application/json
      description: 管理员逻辑删除用户，需 admin 角色
      parameters:
      - description: 删除参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_ai-comic-generator_server_internal_model.DeleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 业务错误
          schema:
            $ref: '#/definitions/github_com_ai-comic-generator_server_internal_common.BaseResponse'
      security:
      - SessionCookie: []
      summary: 删除用户
      tags:
      - 用户管理
  /user/info:
    get:
      description: 从 Session 读取当前登录用户信息，含配额等
      produces:
      - application/json
      responses:
        "200":
          description: 未登录
          schema:
            $ref: '#/definitions/github_com_ai-comic-generator_server_internal_common.BaseResponse'
      security:
      - SessionCookie: []
      summary: 获取当前登录用户
      tags:
      - 用户
  /user/login:
    post:
      consumes:
      - application/json
      description: 登录成功后写入 Session Cookie（session），后续受保护接口需携带
      parameters:
      - description: 登录参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_ai-comic-generator_server_internal_model.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 业务错误（账号或密码错误等）
          schema:
            $ref: '#/definitions/github_com_ai-comic-generator_server_internal_common.BaseResponse'
      summary: 用户登录
      tags:
      - 用户
  /user/logout:
    post:
      description: 清除 Session，退出登录
      produces:
      - application/json
      responses:
        "200":
          description: 业务错误
          schema:
            $ref: '#/definitions/github_com_ai-comic-generator_server_internal_common.BaseResponse'
      security:
      - SessionCookie: []
      summary: 用户注销
      tags:
      - 用户
  /user/page/vo:
    post:
      consumes:
      - application/json
      description: 管理员分页查询用户，支持多条件筛选与排序，需 admin 角色
      parameters:
      - description: 查询参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_ai-comic-generator_server_internal_model.QueryUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 业务错误
          schema:
            $ref: '#/definitions/github_com_ai-comic-generator_server_internal_common.BaseResponse'
      security:
      - SessionCookie: []
      summary: 分页查询用户列表
      tags:
      - 用户管理
  /user/password/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: 原密码与新密码
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_ai-comic-generator_server_internal_model.UpdatePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ai-comic-generator_server_internal_common.BaseResponse'
      security:
      - SessionCookie: []
      summary: 修改密码
      tags:
      - 用户
  /user/profile/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: 昵称/头像/简介
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_ai-comic-generator_server_internal_model.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ai-comic-generator_server_internal_common.BaseResponse'
      security:
      - SessionCookie: []
      summary: 更新个人资料
      tags:
      - 用户
  /user/register:
    post:
      consumes:
      - application/json
      description: 注册新用户，账号至少 4 位，密码至少 8 位
      parameters:
      - description: 注册参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_ai-comic-generator_server_internal_model.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 业务错误（参数错误、账号重复等）
          schema:
            $ref: '#/definitions/github_com_ai-comic-generator_server_internal_common.BaseResponse'
      summary: 用户注册
      tags:
      - 用户
  /user/update:
    post:
      consumes:
      - application/json
      description: 管理员更新用户信息，需 admin 角色
      parameters:
      - description: 更新参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_ai-comic-generator_server_internal_model.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 业务错误
          schema:
            $ref: '#/definitions/github_com_ai-comic-generator_server_internal_common.BaseResponse'
      security:
      - SessionCookie: []
      summary: 更新用户
      tags:
      - 用户管理
schemes:
- http
- https
securityDefinitions:
  SessionCookie:
    in: cookie
    name: session
    type: apiKey
swagger: "2.0"